+++ src/main/java/de/maxhenkel/plane/ServerConfig.java
@@ -20,6 +20,10 @@ public class ServerConfig extends ConfigBase {
     public final ModConfigSpec.IntValue planeFuelCapacity;
     public final ModConfigSpec.IntValue cargoPlaneFuelCapacity;
     public final ModConfigSpec.IntValue transporterPlaneFuelCapacity;
+    public final ModConfigSpec.DoubleValue bushPlaneBaseFuelUsage;
+    public final ModConfigSpec.DoubleValue planeBaseFuelUsage;
+    public final ModConfigSpec.DoubleValue cargoPlaneBaseFuelUsage;
+    public final ModConfigSpec.DoubleValue transporterPlaneBaseFuelUsage;
 
 
     private final ModConfigSpec.ConfigValue<List<? extends String>> validFuelsSpec;
@@ -33,7 +37,12 @@ public class ServerConfig extends ConfigBase {
         bushPlaneFuelCapacity = builder.worldRestart().defineInRange("plane.bush_plane.fuel_capacity", 8_000, 1000, Integer.MAX_VALUE);
         planeFuelCapacity = builder.worldRestart().defineInRange("plane.plane.fuel_capacity", 10_000, 1000, Integer.MAX_VALUE);
         cargoPlaneFuelCapacity = builder.worldRestart().defineInRange("plane.cargo.fuel_capacity", 16_000, 1000, Integer.MAX_VALUE);
-        transporterPlaneFuelCapacity = builder.worldRestart().defineInRange("plane.transporter.fuel", 16_000, 1000, Integer.MAX_VALUE);
+        transporterPlaneFuelCapacity = builder.worldRestart().defineInRange("plane.transporter.fuel_capacity", 16_000, 1000, Integer.MAX_VALUE);
+
+        bushPlaneBaseFuelUsage = builder.worldRestart().defineInRange("plane.bush_plane.base_fuel_usage", 0.75D, 0.01D, 16D);
+        planeBaseFuelUsage = builder.worldRestart().defineInRange("plane.plane.base_fuel_usage", 0.8D, 0.01D, 16D);
+        cargoPlaneBaseFuelUsage = builder.worldRestart().defineInRange("plane.cargo.base_fuel_usage", 1.05D, 0.01D, 16D);
+        transporterPlaneBaseFuelUsage = builder.worldRestart().defineInRange("plane.transporter.base_fuel_usage", 0.95F, 0.01D, 16D);
 
         validFuelsSpec = builder.worldRestart().defineList("plane.valid_fuels", () -> List.of("#car:gas_station"), () -> "", Objects::nonNull);
     }