+++ src/main/java/de/maxhenkel/plane/entity/EntityPlane.java
@@ -66,7 +66,7 @@ public class EntityPlane extends EntityPlaneBase {
 
     @Override
     protected float getBaseFuelUsage() {
-        return 0.75F;
+        return Main.SERVER_CONFIG.planeBaseFuelUsage.get().floatValue();
     }
 
     @Override